requirements.txt
streamlit>=1.35.0
openai>=1.30.0
pydantic>=2.7.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
graphviz>=0.20.3
plotly>=5.22.0
